+++ crates/biome_js_analyze/src/lint/nursery.rs
@@ -17,6 +17,7 @@ pub mod no_useless_string_concat;
 pub mod no_useless_undefined_initialization;
 pub mod use_array_literals;
 pub mod use_consistent_builtin_instantiation;
+pub mod use_date_now;
 pub mod use_default_switch_clause;
 pub mod use_explicit_length_check;
 pub mod use_focusable_interactive;
@@ -47,6 +48,7 @@ declare_group! {
             self :: no_useless_undefined_initialization :: NoUselessUndefinedInitialization ,
             self :: use_array_literals :: UseArrayLiterals ,
             self :: use_consistent_builtin_instantiation :: UseConsistentBuiltinInstantiation ,
+            self :: use_date_now :: UseDateNow ,
             self :: use_default_switch_clause :: UseDefaultSwitchClause ,
             self :: use_explicit_length_check :: UseExplicitLengthCheck ,
             self :: use_focusable_interactive :: UseFocusableInteractive ,